requirements.txt
aiogram>=3.0.0
gspread
google-auth
python-docx
python-dotenv
google-genai
textract==1.6.3